refactor(package): reorganize module structure
