+++ demo_exam.py
@@ -1,5 +1,5 @@
-from examgen.gui.dialogs import ExamConfigDialog
-from examgen.gui.widgets import start_exam
+from examgen.gui.dialogs.question_dialog import ExamConfigDialog
+from examgen.gui.widgets.option_table import start_exam
 from PySide6.QtWidgets import QApplication
 import sys
