+++ pyproject.toml
@@ -8,5 +8,8 @@ version = "0.1.0"
 
 [tool.setuptools]
 package-dir = {"" = "src"}
-packages = ["examgen"]
+
+[tool.setuptools.packages.find]
+where = ["src"]
+include = ["examgen*"]
