+++ src/examgen/gui/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
-from .widgets import ExamDialog, start_exam
+from .widgets.option_table import ExamDialog, start_exam
 
 __all__ = ["ExamDialog", "start_exam"]
