+++ src/examgen/gui/dialogs/question_dialog.py
@@ -3,8 +3,7 @@ from pathlib import Path
 from typing import List, Optional
 
 from PySide6.QtCore import Qt, QSize
-from PySide6.QtGui import QIcon, QTextOption, QColor
-from PySide6.QtWidgets import QLineEdit
+from PySide6.QtGui import QIcon, QTextOption
 from PySide6.QtWidgets import (
     QCheckBox,
     QComboBox,
@@ -28,26 +27,18 @@ from PySide6.QtWidgets import (
     QStyledItemDelegate,
     QStyleOptionViewItem,
     QCompleter,
+    QLineEdit,
 )
 
-from examgen import models as m
-from examgen.models import SessionLocal
-
-from examgen.services.exam_service import (
-    ExamConfig,
-    SelectorTypeEnum,
-    Attempt,
-)
-from sqlalchemy.orm import selectinload, joinedload
+from examgen.core import models as m
+from examgen.core.models import SessionLocal
+from examgen.core.services.exam_service import ExamConfig, SelectorTypeEnum
 
 DB_PATH = Path("examgen.db")
 MAX_CHARS = 3000
 MIN_ROWS = 4
 
 
-# ------------------------------------------------------------------
-# Delegate that wraps text even without spaces
-# ------------------------------------------------------------------
 class WrapAnywhereDelegate(QStyledItemDelegate):
     _flags = Qt.TextWordWrap | Qt.TextWrapAnywhere
 
@@ -65,22 +56,16 @@ class WrapAnywhereDelegate(QStyledItemDelegate):
     def paint(self, painter, option, index):  # type: ignore[override]
         self.initStyleOption(option, index)
         painter.save()
-        painter.drawText(
-            option.rect, self._flags, str(index.data(Qt.DisplayRole) or "")
-        )
+        painter.drawText(option.rect, self._flags, str(index.data(Qt.DisplayRole) or ""))
         painter.restore()
 
 
-# ------------------------------------------------------------------
-# OptionTable
-# ------------------------------------------------------------------
 class OptionTable(QTableWidget):
     HEADER = ["", "Opción", "Correcta", "Explicación", ""]
 
     def __init__(self, parent: QWidget | None = None):
         super().__init__(MIN_ROWS, 5, parent)
 
-        # header setup
         self.setHorizontalHeaderLabels(self.HEADER)
         hh = self.horizontalHeader()
         hh.setSectionResizeMode(0, QHeaderView.ResizeToContents)
@@ -91,9 +76,7 @@ class OptionTable(QTableWidget):
         hh.setStretchLastSection(False)
 
         self.verticalHeader().setVisible(False)
-        self.verticalHeader().setDefaultSectionSize(
-            self.fontMetrics().lineSpacing() + 6
-        )
+        self.verticalHeader().setDefaultSectionSize(self.fontMetrics().lineSpacing() + 6)
         self.setWordWrap(True)
 
         wrap = WrapAnywhereDelegate(self)
@@ -107,8 +90,7 @@ class OptionTable(QTableWidget):
 
         self.cellChanged.connect(self._auto_height)
 
-    # ------------ row helpers -----------------
-    def _init_row(self, row: int):
+    def _init_row(self, row: int) -> None:
         letter = QTableWidgetItem(chr(ord("a") + row))
         letter.setFlags(Qt.ItemIsEnabled)
         letter.setTextAlignment(Qt.AlignCenter)
@@ -132,7 +114,7 @@ class OptionTable(QTableWidget):
         trash.clicked.connect(lambda *_: self._remove_row(row))
         self.setCellWidget(row, 4, trash)
 
-    def add_row(self):
+    def add_row(self) -> None:
         r = self.rowCount()
         self.insertRow(r)
         self._init_row(r)
@@ -140,28 +122,27 @@ class OptionTable(QTableWidget):
         self._refresh_delete_buttons()
         self.resizeRowsToContents()
 
-    def _remove_row(self, row: int):
+    def _remove_row(self, row: int) -> None:
         if self.rowCount() > MIN_ROWS:
             self.removeRow(row)
             self._refresh_letters()
             self._refresh_delete_buttons()
             self.resizeRowsToContents()
 
-    def _refresh_letters(self):
+    def _refresh_letters(self) -> None:
         for r in range(self.rowCount()):
             self.item(r, 0).setText(chr(ord("a") + r))
 
-    def _refresh_delete_buttons(self):
+    def _refresh_delete_buttons(self) -> None:
         allow = self.rowCount() > MIN_ROWS
         for r in range(self.rowCount()):
-            btn: QToolButton | None = self.cellWidget(r, 4).findChild(QToolButton)  # type: ignore
+            btn: QToolButton | None = self.cellWidget(r, 4).findChild(QToolButton)
             if btn:
                 btn.setEnabled(allow)
 
-    def _auto_height(self, row: int, _col: int):
+    def _auto_height(self, row: int, _col: int) -> None:
         self.resizeRowToContents(row)
 
-    # ------------ collect ---------------------
     def collect(self) -> tuple[List[m.AnswerOption], int]:
         opts: List[m.AnswerOption] = []
         correct = 0
@@ -172,7 +153,7 @@ class OptionTable(QTableWidget):
             text = text[:MAX_CHARS]
             answer = None
             expl = self.item(r, 3).text().strip()[:MAX_CHARS]
-            is_corr = self.cellWidget(r, 2).findChild(QCheckBox).isChecked()  # type: ignore
+            is_corr = self.cellWidget(r, 2).findChild(QCheckBox).isChecked()
             if is_corr:
                 correct += 1
             opts.append(
@@ -186,9 +167,6 @@ class OptionTable(QTableWidget):
         return opts, correct
 
 
-# ------------------------------------------------------------------
-# ExamConfigDialog
-# ------------------------------------------------------------------
 class ExamConfigDialog(QDialog):
     """Dialog to configure exam parameters."""
 
@@ -200,10 +178,8 @@ class ExamConfigDialog(QDialog):
         self.cb_subject = QComboBox()
         self.cb_subject.setEditable(True)
         self.spin_time = QSpinBox(minimum=1, maximum=999, value=90)
-
         self.spin_questions = QSpinBox(minimum=1, maximum=999, value=60)
 
-        # match widths for spin boxes
         width = self.spin_time.sizeHint().width()
         self.spin_time.setFixedWidth(width)
         self.spin_questions.setFixedWidth(width)
@@ -254,9 +230,7 @@ class ExamConfigDialog(QDialog):
         self.spin_questions.valueChanged.connect(self._update_ok_state)
         self.spin_time.valueChanged.connect(self._update_ok_state)
 
-    # ---------------- helpers -----------------
     def _load_subjects(self) -> None:
-        """Populate subject combo from the database."""
         with SessionLocal() as s:
             subjects = s.query(m.Subject).order_by(m.Subject.name).all()
 
@@ -265,7 +239,6 @@ class ExamConfigDialog(QDialog):
             self.cb_subject.addItem(subj.name, subj.id)
 
         if self.cb_subject.isEditable():
-            # ensure editable combo before attaching a completer
             self.cb_subject.setEditable(True)
             completer = QCompleter(self.cb_subject.model(), self)
             completer.setCaseSensitivity(Qt.CaseInsensitive)
@@ -292,12 +265,9 @@ class ExamConfigDialog(QDialog):
         ok_enabled = subject_ok and selector_ok and questions_ok and time_ok
         self.btn_ok.setEnabled(ok_enabled)
 
-    # ---------------- accept -------------------
     def accept(self) -> None:  # type: ignore[override]
         selector = (
-            SelectorTypeEnum.ALEATORIO
-            if self.rb_random.isChecked()
-            else SelectorTypeEnum.ERRORES
+            SelectorTypeEnum.ALEATORIO if self.rb_random.isChecked() else SelectorTypeEnum.ERRORES
         )
         self.config = ExamConfig(
             exam_id=0,
@@ -310,16 +280,12 @@ class ExamConfigDialog(QDialog):
         )
         super().accept()
 
-    # ---------------- convenience -------------
     @classmethod
     def get_config(cls, parent: QWidget | None = None) -> Optional[ExamConfig]:
         dlg = cls(parent)
         return dlg.config if dlg.exec() == cls.Accepted else None
 
 
-# ------------------------------------------------------------------
-# QuestionDialog
-# ------------------------------------------------------------------
 class QuestionDialog(QDialog):
     def __init__(self, parent: QWidget | None = None, *, db_path: Path = DB_PATH):
         super().__init__(parent)
@@ -327,7 +293,6 @@ class QuestionDialog(QDialog):
         self.setWindowTitle("Nueva pregunta – MCQ")
         self.resize(1920, 1080)
 
-        # subject / reference
         self.cb_subject = QComboBox(editable=True, fixedWidth=500)
         self.le_reference = QLineEdit()
         self.le_reference.setPlaceholderText("ej.: exam_1025")
@@ -341,7 +306,6 @@ class QuestionDialog(QDialog):
         w_top = QWidget()
         w_top.setLayout(top)
 
-        # prompt
         self.prompt = QPlainTextEdit()
         self.prompt.setWordWrapMode(QTextOption.WrapMode.WordWrap)
         self.counter = QLabel("0/3000", alignment=Qt.AlignRight)
@@ -352,7 +316,6 @@ class QuestionDialog(QDialog):
         hp.setContentsMargins(0, 0, 0, 0)
         hp.addWidget(self.prompt)
 
-        # table options
         self.table = OptionTable(self)
         add_btn = QPushButton("+", clicked=self.table.add_row, fixedSize=QSize(30, 30))
 
@@ -376,31 +339,26 @@ class QuestionDialog(QDialog):
         root.addWidget(buttons)
         self._load_subjects()
 
-    # ---------------- helpers -----------------
-    def _update_counter(self):
+    def _update_counter(self) -> None:
         txt = self.prompt.toPlainText()
         if len(txt) > MAX_CHARS:
             self.prompt.setPlainText(txt[:MAX_CHARS])
             self.prompt.moveCursor(self.prompt.textCursor().End)
         self.counter.setText(f"{len(self.prompt.toPlainText())}/{MAX_CHARS}")
 
-    def _load_subjects(self):
-        """Carga materias existentes en el combo de materia."""
+    def _load_subjects(self) -> None:
         with m.Session(m.get_engine(self.db_path)) as s:
             names = sorted(sub.name for sub in s.query(m.Subject).all())
         self.cb_subject.addItems(names)
         self.cb_subject.setPlaceholderText("Seleccione / escriba…")
 
-    # ---------------- guardar -----------------
-    def accept(self):  # type: ignore[override]  noqa: D401
+    def accept(self) -> None:  # type: ignore[override]
         subj = self.cb_subject.currentText().strip()
         ref = self.le_reference.text().strip()
         prompt_txt = self.prompt.toPlainText().strip()
 
         if not subj or not prompt_txt:
-            QMessageBox.warning(
-                self, "Datos incompletos", "Materia y enunciado obligatorios."
-            )
+            QMessageBox.warning(self, "Datos incompletos", "Materia y enunciado obligatorios.")
             return
 
         options, correct = self.table.collect()
@@ -414,263 +372,12 @@ class QuestionDialog(QDialog):
 
         engine = m.get_engine(self.db_path)
         with m.Session(engine) as s:
-            subj_obj = s.query(m.Subject).filter_by(name=subj).first() or m.Subject(
-                name=subj
-            )
+            subj_obj = s.query(m.Subject).filter_by(name=subj).first() or m.Subject(name=subj)
 
-            q = m.MCQQuestion(
-                prompt=prompt_txt,
-                subject=subj_obj,
-                reference=ref or None,
-            )
+            q = m.MCQQuestion(prompt=prompt_txt, subject=subj_obj, reference=ref or None)
             q.options = options
             s.add(q)
             s.commit()
 
-        QMessageBox.information(
-            self, "Pregunta guardada", "La pregunta se ha almacenado correctamente."
-        )
+        QMessageBox.information(self, "Pregunta guardada", "La pregunta se ha almacenado correctamente.")
         super().accept()
-
-
-class ResultsDialog(QDialog):
-    """Show exam results with per-question breakdown."""
-
-    def __init__(self, attempt: Attempt, parent: QWidget | None = None) -> None:
-        # 'attempt' viene ya evaluado y con sesión viva
-        super().__init__(parent)
-        self.attempt = attempt
-
-        self.setWindowTitle("\u00a1Resultados del examen!")
-
-        title = QLabel("\u00a1Resultados del examen!", alignment=Qt.AlignCenter)
-        score = attempt.score or 0
-        total = len(attempt.questions)
-        pct = round((score / total) * 100) if total else 0
-        summary = QLabel(
-            f"Puntuaci\u00f3n: {score} / {total}   ({pct} %)",
-            alignment=Qt.AlignCenter,
-        )
-
-        self.table = QTableWidget(len(attempt.questions), 4, self)
-        self.table.setHorizontalHeaderLabels(["#", "Pregunta", "Tu resp.", "Correcta"])
-        self.table.horizontalHeader().setStretchLastSection(True)
-        self.table.verticalHeader().setVisible(False)
-        self.table.setTextElideMode(Qt.ElideRight)
-        self.table.setMinimumWidth(700)
-
-        correct_bg = QColor("#5af16a")
-        correct_fg = QColor("#000000")
-        wrong_bg = QColor("#ff6b6b")
-        wrong_fg = QColor("#ffffff")
-
-        for row, aq in enumerate(attempt.questions, start=0):
-            qitem = QTableWidgetItem(str(row + 1))
-            qitem.setFlags(Qt.ItemIsEnabled)
-            qitem.setTextAlignment(Qt.AlignCenter)
-            self.table.setItem(row, 0, qitem)
-
-            prompt = aq.question.prompt
-            pitem = QTableWidgetItem(prompt)
-            pitem.setFlags(Qt.ItemIsEnabled)
-            pitem.setToolTip(prompt)
-            self.table.setItem(row, 1, pitem)
-
-            sel_text = aq.selected_option or ""
-            sitem = QTableWidgetItem(sel_text)
-            sitem.setFlags(Qt.ItemIsEnabled)
-            sitem.setToolTip(sel_text)
-            self.table.setItem(row, 2, sitem)
-
-            corr_text = next((o.text for o in aq.question.options if o.is_correct), "")
-            citem = QTableWidgetItem(corr_text)
-            citem.setFlags(Qt.ItemIsEnabled)
-            citem.setToolTip(corr_text)
-            self.table.setItem(row, 3, citem)
-
-            if aq.is_correct:
-                bg = correct_bg
-                fg = correct_fg
-            else:
-                bg = wrong_bg
-                fg = wrong_fg
-            for c in range(4):
-                cell = self.table.item(row, c)
-                cell.setBackground(bg)
-                cell.setForeground(fg)
-
-        buttons = QDialogButtonBox(QDialogButtonBox.Close, parent=self)
-        buttons.rejected.connect(self.reject)
-
-        root = QVBoxLayout(self)
-        root.addWidget(title)
-        root.addWidget(summary)
-        root.addWidget(self.table)
-        root.addWidget(buttons, alignment=Qt.AlignCenter)
-
-        self.resize(self.width() + 120, self.height())
-
-    @classmethod
-    def show_for_attempt(cls, attempt: Attempt, parent: QWidget | None = None) -> None:
-        session = SessionLocal()
-        attempt_db = (
-            session.query(m.Attempt)
-            .options(
-                selectinload(m.Attempt.questions)
-                .joinedload(
-                    m.AttemptQuestion.question.of_type(m.MCQQuestion)
-                )
-                .selectinload(m.MCQQuestion.options)
-            )
-            .get(attempt.id)
-        )
-        if attempt_db is None:
-            QMessageBox.critical(parent, "Error", "Intento no encontrado en BD")
-            session.close()
-            return
-
-        dlg = cls(attempt_db, parent)
-        dlg._session = session
-        dlg.exec()
-
-    def reject(self) -> None:  # type: ignore[override]
-        if hasattr(self, "_session"):
-            self._session.close()
-        super().reject()
-
-
-class AttemptsHistoryDialog(QDialog):
-    def __init__(self, parent: QWidget | None = None):
-        super().__init__(parent)
-        self.setWindowTitle("Historial de pruebas")
-
-        cols = [
-            "Materia",
-            "Inicio",
-            "Duración",
-            "Preguntas",
-            "Correctas",
-            "%",
-            "",
-        ]
-        self.table = QTableWidget(0, len(cols), self)
-        self.table.setHorizontalHeaderLabels(cols)
-        self.table.horizontalHeader().setStretchLastSection(False)
-        self.table.verticalHeader().setVisible(False)
-
-        self._session = SessionLocal()
-
-        self.btn_clear = QPushButton("Borrar todo")
-        self.btn_close = QPushButton("Cerrar")
-        self.btn_clear.clicked.connect(self._clear_all)
-        self.btn_close.clicked.connect(self.reject)
-
-        root = QVBoxLayout(self)
-        root.addWidget(self.table)
-        footer = QHBoxLayout()
-        footer.addStretch(1)
-        footer.addWidget(self.btn_clear)
-        footer.addWidget(self.btn_close)
-        root.addLayout(footer)
-
-        self._reload_table()
-
-    def _reload_table(self) -> None:
-        attempts = (
-            self._session.query(m.Attempt)
-            .options(selectinload(m.Attempt.questions))
-            .order_by(m.Attempt.started_at.desc())
-            .all()
-        )
-
-        fmt = "%d/%m/%Y %H:%M"
-        self.table.setRowCount(len(attempts))
-        for row, at in enumerate(attempts):
-            start = at.started_at.strftime(fmt) if at.started_at else "-"
-            if at.ended_at:
-                secs = int((at.ended_at - at.started_at).total_seconds())
-                dur_txt = f"{secs//60}:{secs%60:02d} min"
-            else:
-                dur_txt = "-"
-
-            total_q = len(at.questions)
-            corr = at.score or 0
-            pct = round((corr / total_q) * 100) if total_q else 0
-
-            vals = [at.subject, start, dur_txt, str(total_q), str(corr), f"{pct} %"]
-            for col, val in enumerate(vals):
-                item = QTableWidgetItem(val)
-                item.setFlags(Qt.ItemIsEnabled)
-                if col == 0:
-                    item.setTextAlignment(Qt.AlignLeft | Qt.AlignVCenter)
-                else:
-                    item.setTextAlignment(Qt.AlignCenter)
-                self.table.setItem(row, col, item)
-
-            del_btn = QToolButton(text="🗑️")
-            del_btn.setAutoRaise(True)
-            del_btn.setCursor(Qt.PointingHandCursor)
-            del_btn.clicked.connect(lambda _, aid=at.id: self._delete_attempt(aid))
-            self.table.setCellWidget(row, 6, del_btn)
-
-        for c in range(self.table.columnCount()):
-            self.table.resizeColumnToContents(c)
-            w_head = self.table.horizontalHeader().sectionSizeHint(c)
-            w_cells = max(self.table.sizeHintForColumn(c), w_head)
-            self.table.setColumnWidth(c, w_cells + 12)
-
-        total_w = sum(self.table.columnWidth(c) for c in range(self.table.columnCount()))
-        total_w += self.table.verticalHeader().width() + 40
-        self.table.setMinimumWidth(total_w)
-        self.resize(total_w, self.height())
-
-    def _delete_attempt(self, aid: int) -> None:
-        if (
-            QMessageBox.question(
-                self,
-                "Borrar examen",
-                "¿Seguro que deseas borrar este examen?",
-                QMessageBox.Yes | QMessageBox.No,
-                QMessageBox.No,
-            )
-            != QMessageBox.Yes
-        ):
-            return
-        with SessionLocal() as s:
-            s.query(m.Attempt).filter_by(id=aid).delete()
-            s.commit()
-        self._reload_table()
-
-    def _clear_all(self) -> None:
-        if (
-            QMessageBox.question(
-                self,
-                "Borrar historial",
-                "¿Seguro que deseas borrar todo el historial?",
-                QMessageBox.Yes | QMessageBox.No,
-                QMessageBox.No,
-            )
-            != QMessageBox.Yes
-        ):
-            return
-        with SessionLocal() as s:
-            s.query(m.Attempt).delete()
-            s.commit()
-        self._reload_table()
-
-    def reject(self) -> None:  # type: ignore[override]
-        if hasattr(self, "_session"):
-            self._session.close()
-        super().reject()
-
-
-if __name__ == "__main__":
-    import sys
-    from PySide6.QtWidgets import QApplication
-    from examgen.models import SessionLocal
-
-    app = QApplication(sys.argv)
-    with SessionLocal() as s:
-        attempt = s.query(Attempt).order_by(Attempt.id.desc()).first()
-    if attempt:
-        ResultsDialog.show_for_attempt(attempt)
